wip
tests that check if the delim assignment treats standalone right,

i.e. if the assignment is the only thing on the line, omit the line.
and maybe strip leading trailing whitespace too though I'm not sure

are failing. I would need to emit a node in the ast for the assignment
and then strip whitespace like comments do.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>